Update README.md
We are creating a new fork to prevent confusion between OpenID 2.0 and OpenID Connect. This fork always implemented the latter, also known as OIDC. The new fork clearifies this by using the name Mailu-OIDC. We decided against renaming to ensure historic references to this fork are still valid and working.